{refactor} fixed typo in environment import in app.module.ts file and main.ts file from evironment to environment
