+++ apps/conv-learning-manager/src/main.ts
@@ -2,7 +2,7 @@ import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
 import { AppModule } from './app/app.module';
-import { environment } from '../src/environments/evironment';
+import { environment } from '../src/environments/environment';
 
 if (environment.production) {
   enableProdMode();
